+++ lib/flutter_saucenao_img_search.dart
@@ -3,3 +3,4 @@ library flutter_saucenao_img_search;
 export 'src/image_searcher.dart';
 export 'src/searcher_config.dart';
 export 'src/search_result.dart';
+export 'src/exceptions.dart';