+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.1.3
 - Update `README.md`
+- Fix the issue that the Exceptions is not visible
 
 ## 0.1.2
 - Update `README.md`